python/pyproject.toml
[tool.poetry]
name = "alexa-for-docs"
version = "0.1.0"
description = ""
authors = ["elenacramer <elenascramer@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
onnxruntime = "^1.6.0"
onnxruntime-tools = "^1.5.2"
azure-functions = "^1.5.0"
farm-haystack = "^0.6.0"
pandas = "^1.2.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
notebook = "^6.1.5"
sentencepiece = "^0.1.94"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
